🚁 "Extracted axios instance to api_url.ts, updated schema types to use int() method, and removed redundant imports."

+++ frontend/web/app/src/schemas/account.ts
@@ -16,8 +16,8 @@ const AccountIn = z.object(
 
 const AccountOut = AccountIn.merge(
   z.object({
-    id: z.number(),
-    userId: z.number()
+    id: z.number().int(),
+    userId: z.number().int()
   })
 )
